refactor: move requests to own package
