+++ src/main/java/se/ltu/eduo/controller/CollectionController.java
@@ -8,6 +8,10 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import se.ltu.eduo.dto.*;
+import se.ltu.eduo.dto.request.CreateCollectionRequest;
+import se.ltu.eduo.dto.request.CreateGenerationRequest;
+import se.ltu.eduo.dto.request.UpdateProjectRequest;
+import se.ltu.eduo.dto.request.UpdateQuizRequest;
 import se.ltu.eduo.mapper.CollectionMapper;
 import se.ltu.eduo.mapper.GenerationMapper;
 import se.ltu.eduo.mapper.QuizMapper;
