+++ src/main/java/se/ltu/eduo/controller/UserController.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import se.ltu.eduo.dto.LoginRequest;
-import se.ltu.eduo.dto.RegisterRequest;
+import se.ltu.eduo.dto.request.LoginRequest;
+import se.ltu.eduo.dto.request.RegisterRequest;
 import se.ltu.eduo.dto.UserDto;
 import se.ltu.eduo.exception.UsernameAlreadyExistsException;
 import se.ltu.eduo.mapper.UserMapper;
