--- src/main/java/se/ltu/eduo/dto/CreateGenerationRequest.java
@@ -1,6 +1,0 @@
-package se.ltu.eduo.dto;
-
-import java.util.List;
-import java.util.UUID;
-
-public record CreateGenerationRequest(List<UUID> sourceMaterialIds) {}
