+++ src/main/java/se/ltu/eduo/dto/request/CreateCollectionRequest.java
@@ -1,3 +1,3 @@
-package se.ltu.eduo.dto;
+package se.ltu.eduo.dto.request;
 
 public record CreateCollectionRequest(Integer userId, String name) {}
