+++ src/main/java/se/ltu/eduo/dto/request/LoginRequest.java
@@ -1,3 +1,3 @@
-package se.ltu.eduo.dto;
+package se.ltu.eduo.dto.request;
 
 public record LoginRequest(String username, String password) {}
