+++ src/main/java/se/ltu/eduo/dto/request/RegisterRequest.java
@@ -1,3 +1,3 @@
-package se.ltu.eduo.dto;
+package se.ltu.eduo.dto.request;
 
 public record RegisterRequest(String firstName, String lastName, String username, String password) {}
