+++ src/main/java/se/ltu/eduo/dto/request/UpdateProjectRequest.java
@@ -1,3 +1,3 @@
-package se.ltu.eduo.dto;
+package se.ltu.eduo.dto.request;
 
 public record UpdateProjectRequest(String name) {}
