+++ src/main/java/se/ltu/eduo/dto/request/UpdateQuizRequest.java
@@ -1,3 +1,3 @@
-package se.ltu.eduo.dto;
+package se.ltu.eduo.dto.request;
 
 public record UpdateQuizRequest(String name, String rawContent) {}
